requirements.txt
requests>=2.31.0
opensearch-py>=2.3.0
pyyaml>=6.0
faker>=18.0.0
python-dateutil>=2.8.2
schedule>=1.2.0
requests-aws4auth>=1.2.0
pydantic>=2.0.0
opensearch-py>=2.3.0
networkx>=2.6.3,<3.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
strands-agents
strands-agents-tools